Ensure that all utilities still work
