+++ lib/create.js
@@ -6,8 +6,7 @@ const path = require('path')
 const createDb = (file) => flume(log(file, { codec: codec.json }))
 
 const paths = {
-  db: path.join(__dirname, 'db'),
-  tmp: '/tmp/db'
+  db: path.join(__dirname, './.test-db')
 }
 var a = createDb(paths.db)
 const messages = [
