+++ package.json
@@ -15,10 +15,8 @@
     "flumedb": "github:flumedb/flumedb#del-only",
     "flumelog-offset": "github:flumedb/flumelog-offset#del",
     "mutant": "^3.22.1",
-    "mv": "^2.1.1",
     "patchcore": "^2.1.0",
     "pull-stream": "^3.6.9",
-    "rimraf": "^2.6.2",
     "ssb-ref": "^2.13.8",
     "yargs": "^12.0.5"
   }
